Added mechanism to have callables which mutate the environment.
Right now 'defun' is built into the grammar, but this feels awkward.
So instead I'm trying to define a new type of callable which can
mutate its environment, and then let 'defun' get handled like a
normal function lookup during evaluation. This will simplify parsing
and provide a new type of callable that might be valuable in the future.